+++ src/annotations/annotations.ts
@@ -4,7 +4,6 @@ import { DiffWithCommand, OpenCommitInRemoteCommand, ShowQuickCommitDetailsComma
 import { IThemeConfig, themeDefaults } from '../configuration';
 import { GlyphChars } from '../constants';
 import { CommitFormatter, GitCommit, GitDiffChunkLine, GitService, GitUri, ICommitFormatOptions } from '../gitService';
-const Datauri = require('datauri');
 
 interface IHeatmapConfig {
     enabled: boolean;
@@ -25,25 +24,6 @@ interface IRenderOptions {
 export const endOfLineIndex = 1000000;
 const escapeMarkdownRegEx = /[`\>\#\*\_\-\+\.]/g;
 // const sampleMarkdown = '## message `not code` *not important* _no underline_ \n> don\'t quote me \n- don\'t list me \n+ don\'t list me \n1. don\'t list me \nnot h1 \n=== \nnot h2 \n---\n***\n---\n___';
-
-const linkIconSvg = `<?xml version="1.0" encoding="utf-8"?>
-<svg width="12" height="18" version="1.1" xmlns="http://www.w3.org/2000/svg">
-    <path fill="\${color}" d="m11,14l1,0l0,3c0,0.55 -0.45,1 -1,1l-10,0c-0.55,0 -1,-0.45 -1,-1l0,-10c0,-0.55 0.45,-1 1,-1l3,0l0,1l-3,0l0,10l10,0l0,-3l0,0zm-5,-8l2.25,2.25l-3.25,3.25l1.5,1.5l3.25,-3.25l2.25,2.25l0,-6l-6,0l0,0z" />
-</svg>`;
-
-const themeForegroundColor = '#a0a0a0';
-let linkIconDataUri: string | undefined;
-
-function getLinkIconDataUri(foregroundColor: string): string {
-    if (linkIconDataUri === undefined || foregroundColor !== themeForegroundColor) {
-        const datauri = new Datauri();
-        datauri.format('.svg', Strings.interpolate(linkIconSvg, { color: foregroundColor }));
-        linkIconDataUri = datauri.content;
-        foregroundColor = themeForegroundColor;
-    }
-
-    return linkIconDataUri!;
-}
 
 export class Annotations {
 
@@ -85,7 +65,7 @@ export class Annotations {
         }
 
         const openInRemoteCommand = hasRemotes
-            ? `${'&nbsp;'.repeat(3)} [![](${getLinkIconDataUri(themeForegroundColor)})](${OpenCommitInRemoteCommand.getMarkdownCommandArgs(commit.sha)} "Open in Remote")`
+            ? `${'&nbsp;'.repeat(3)} [\`${GlyphChars.ArrowUpRight}\`](${OpenCommitInRemoteCommand.getMarkdownCommandArgs(commit.sha)} "Open in Remote")`
             : '';
 
         const markdown = new MarkdownString(`[\`${commit.shortSha}\`](${ShowQuickCommitDetailsCommand.getMarkdownCommandArgs(commit.sha)} "Show Commit Details") &nbsp; __${commit.author}__, ${commit.fromNow()} &nbsp; _(${commit.formatDate(dateFormat)})_ ${openInRemoteCommand} &nbsp; ${message}`);